+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/module/HumanResourceModule.java
@@ -45,6 +45,8 @@ import com.axelor.apps.hr.db.repo.TimesheetTimerHRRepository;
 import com.axelor.apps.hr.db.repo.UserHRRepository;
 import com.axelor.apps.hr.service.app.AppHumanResourceService;
 import com.axelor.apps.hr.service.app.AppHumanResourceServiceImpl;
+import com.axelor.apps.hr.service.app.AppTimesheetService;
+import com.axelor.apps.hr.service.app.AppTimesheetServiceImpl;
 import com.axelor.apps.hr.service.bankorder.BankOrderLineOriginServiceHRImpl;
 import com.axelor.apps.hr.service.bankorder.BankOrderMergeHRServiceImpl;
 import com.axelor.apps.hr.service.bankorder.BankOrderServiceHRImpl;
@@ -117,5 +119,6 @@ public class HumanResourceModule extends AxelorModule {
     bind(PartnerAccountRepository.class).to(PartnerHRRepository.class);
     bind(BankOrderMergeServiceImpl.class).to(BankOrderMergeHRServiceImpl.class);
     bind(TimesheetReportService.class).to(TimesheetReportServiceImpl.class);
+    bind(AppTimesheetService.class).to(AppTimesheetServiceImpl.class);
   }
 }